add test job running; does not work

+++ middlewares/selectOpenErrorCount.sql
@@ -1,5 +1,5 @@
 -- BLOCK select_open_error_count
-SELECT count(*)
+SELECT count(*)::int
 FROM errors AS e
 WHERE
     e.course_id = $course_id
